requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
pip-tools>=7.3.0
pandas>=2.0.0
tenacity>=8.2.0